backend/requirements.txt
Flask>=2.2
flask-cors>=4.0
pandas>=2.0
numpy>=1.25
python-dateutil>=2.8
validate-docbr>=1.10.0
openpyxl>=3.1
pytest>=8.0
